Accept either files or text for drag-and-drop
The use case for this is a bank that doesn't offer CSV downloads, only
Excel (yes, actual Excel files!) with some extra rows at the top.  It
turns out that I can just select the list of transactions in the web
interface, and drag it over to ynab-csv and drop it.